add connection handler
