+++ calvin/runtime/north/actormanager.py
@@ -21,6 +21,7 @@ from calvin.utilities.calvin_callback import CalvinCB
 import calvin.utilities.calvinresponse as response
 from calvin.actor.actor_factory import ActorFactory
 from calvin.actor.actor import ShadowActor
+from calvin.actor.connection_handler import ConnectionHandler
 
 _log = get_logger(__name__)
 
@@ -34,11 +35,12 @@ class ActorManager(object):
 
     """docstring for ActorManager"""
 
-    def __init__(self, node, factory=None):
+    def __init__(self, node, factory=None, connection_handler=None):
         super(ActorManager, self).__init__()
         self.actors = {}
         self.node = node
         self.factory = factory if factory else ActorFactory(node)
+        self.connection_handler = connection_handler if connection_handler else ConnectionHandler(node)
 
     def new(self, actor_type, args, state=None, prev_connections=None, connection_list=None, callback=None,
             signature=None):
@@ -57,7 +59,7 @@ class ActorManager(object):
         _log.debug("class: %s args: %s state: %s, signature: %s" % (actor_type, args, state, signature))
 
         a = self._new(actor_type, args, state, signature)
-        self._setup_connections(a, prev_connections=prev_connections, connection_list=connection_list)
+        self.connection_handler.setup_connections(a, prev_connections=prev_connections, connection_list=connection_list)
 
         if callback:
             callback(status=response.CalvinResponse(True), actor_id=a.id)
@@ -128,7 +130,7 @@ class ActorManager(object):
         _log.analyze(self.node.id, "+ CALL CB", {'actor_id': actor_id, 'node_iter': str(node_iter)})
         # Must call it since the triggers might already have released before cb set
         self._update_requirements_placements(node_iter, actor_id, possible_placements,
-                                 move=move, cb=callback, done=done)
+                                             move=move, cb=callback, done=done)
         _log.analyze(self.node.id, "+ END", {'actor_id': actor_id, 'node_iter': str(node_iter)})
 
     def _update_requirements_placements(self, node_iter, actor_id, possible_placements, done, move=False, cb=None):
@@ -154,8 +156,8 @@ class ActorManager(object):
             delay = 0.0 if actor._collect_placement_counter > actor._collect_placement_last_value + 100 else 0.2
             actor._collect_placement_counter += 1
             actor._collect_placement_cb = async.DelayedCall(delay, self._update_requirements_placements,
-                                                    node_iter, actor_id, possible_placements, done=done,
-                                                     move=move, cb=cb)
+                                                            node_iter, actor_id, possible_placements, done=done,
+                                                            move=move, cb=cb)
             return
         except StopIteration:
             # all possible actor placements derived
@@ -224,67 +226,11 @@ class ActorManager(object):
             # Just pass on new cmd reply if it failed
             self.node.set_local_reply(kwargs['lmsg_id'], reply)
 
-    def _prev_connections_to_connection_list(self, prev_connections):
-        """Convert prev_connection format to connection_list format"""
-        cl = []
-        for in_port_id, out_id in prev_connections['inports'].iteritems():
-            cl.append((self.node.id, in_port_id, out_id[0], out_id[1]))
-        for out_port_id, in_list in prev_connections['outports'].iteritems():
-            for in_id in in_list:
-                cl.append((self.node.id, out_port_id, in_id[0], in_id[1]))
-        return cl
-
-    def _setup_connections(self, actor, prev_connections=None, connection_list=None, callback=None):
-        if prev_connections:
-            # Convert prev_connections to connection_list format
-            connection_list = self._prev_connections_to_connection_list(prev_connections)
-
-        if connection_list:
-            # Migrated actor
-            self.connect(actor.id, connection_list, callback=callback)
-
-    def connect(self, actor_id, connection_list, callback=None):
-        """
-        Reconnecting the ports can be done using a connection_list
-        of tuples (node_id i.e. our id, port_id, peer_node_id, peer_port_id)
-        """
-        if actor_id not in self.actors:
-            return
-
-        peer_port_ids = [c[3] for c in connection_list]
-
-        for node_id, port_id, peer_node_id, peer_port_id in connection_list:
-            self.node.pm.connect(port_id=port_id,
-                                 peer_node_id=peer_node_id,
-                                 peer_port_id=peer_port_id,
-                                 callback=CalvinCB(self._actor_connected,
-                                                   peer_port_id=peer_port_id,
-                                                   actor_id=actor_id,
-                                                   peer_port_ids=peer_port_ids,
-                                                   _callback=callback))
-
-    def _actor_connected(self, status, peer_port_id, actor_id, peer_port_ids, _callback, **kwargs):
-        """ Get called for each of the actor's ports when connecting, but callback should only be called once
-            status: success or not
-            _callback: original callback
-            peer_port_ids: list of port ids kept in context between calls when *changed* by this function,
-                           do not replace it
-        """
-        # Send negative response if not already done it
-        if not status and peer_port_ids:
-            if _callback:
-                del peer_port_ids[:]
-                _callback(status=response.CalvinResponse(False), actor_id=actor_id)
-        if peer_port_id in peer_port_ids:
-            # Remove this port from list
-            peer_port_ids.remove(peer_port_id)
-            # If all ports done send OK
-            if not peer_port_ids:
-                if _callback:
-                    _callback(status=response.CalvinResponse(True), actor_id=actor_id)
-
     def connections(self, actor_id):
-        return self.actors.get(actor_id, None).connections(self.node.id)
+        if actor_id not in self.actors:
+            return []
+
+        return self.connection_handler.connections(self.actors[actor_id])
 
     def dump(self, actor_id):
         actor = self.actors.get(actor_id, None)
@@ -330,4 +276,3 @@ class ActorManager(object):
 
     def list_actors(self):
         return self.actors.keys()
-
